Create README.md
# Palindrome Number Check
A palindrome is a word or phrase that reads the same backward as it does forward. For example, the words "racecar" and "madam" are palindromes. Palindromes can also be numbers, such as 121 and 545.

This C program checks whether a given number is a palindrome. The program first prompts the user to enter a number. Then, it reverses the number and compares it to the original number. If the two numbers are the same, then the program prints The number is a palindrome. Otherwise, the program prints The number is not a palindrome.





## C program to check if a number is a palindrome:

- Prompt the user to enter a number.
- Reverse the number.
- Compare the reversed number to the original number.
- If the two numbers are the same, print "The number is a palindrome." Otherwise, print "The number is not a palindrome."







## Program Flow Chart :

![flowhchart](https://www.enggstudy.com/wp-content/uploads/2022/09/palindrome-number.jpg)
## Usage

## Instalation
To use this program in your local machine , follow these steps:



## Install &  Run Locally

Clone the project

```bash
  git clone https://github.com/coderpappu/problem-solving-By-C.git
```

Go to the project client directory

```bash
  cd Basic Level
```

```bash
  cd Problem-seven
```



Start the Program

```bash
  click the run icon and run the program
```